Added deletion safety measure

+++ src/services/catalog.js
@@ -1,4 +1,5 @@
 const Catalog = require('../models/catalog')
+const Room = require('../models/room')
 
 // Create new catalog
 const catalogNew = async (catalogObject, _id) => {
@@ -30,9 +31,14 @@ const catalogEdit = async (catalogObject) => {
 
 // Delete catalog info
 const catalogDelete = async (catalogObject, _id) => {
-    await Catalog.findOneAndDelete({ _id: catalogObject._id, owner: _id })
-}
+    const rooms = await Room.find({ catalog: catalogObject._id })
 
+    if (rooms.length > 0) {
+        throw new Error('Error: Catalog deletion failed.')
+    } else {
+        await Catalog.findOneAndDelete({ _id: catalogObject._id, owner: _id })
+    }
+}
 
 module.exports = {
     catalogNew,
